requirements.txt
# requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
Pillow>=8.0.0
tqdm>=4.50.0
matplotlib>=3.3.2
einops>=0.3.0